challen1/pyproject.toml
[project]
name = "sitcon-camp-aichallenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.104.1",
    "jinja2==3.1.2",
    "openai==0.28",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "uvicorn==0.24.0",
]
